SCP: Add missing aspects to sim_filepath_parts, move sim_dir_scan to sim_fio
